+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Terraform**
 - Add `testDir` parameter to `init` method.
+- Add `raw` parameter to `output` method.
 
 ### 2.1.2
 **Helm**