+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add `testDir` parameter to `init` method.
 - Add `raw` parameter to `output` method.
 - Add `compactWarn` parameter to `plan` method.
+- Add `rootDir` parameter default value for `providers` method.
+- Fix `rootDir` existence check in `providers` method.
 
 ### 2.1.2
 **Helm**