+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.1.3 (Next)
 **GoSS**
 - Add `logLevel` parameter to `server` method.
+- Remove deprecated format values for `validate` method.
 
 **Packer**
 - Add `recursive` parameter to `fmt` method.